+++ plugins/ucb1x00-raw.c
@@ -23,7 +23,7 @@ static int ucb1x00_read(struct tslib_module_info *inf, struct ts_sample *samp, i
 	ret = read(ts->fd, ucb1x00_evt, sizeof(*ucb1x00_evt) * nr);
 	if(ret > 0) {
 		int nr = ret / sizeof(*ucb1x00_evt);
-		while(ret >= sizeof(*ucb1x00_evt)) {
+		while(ret >= (int)sizeof(*ucb1x00_evt)) {
 			samp->x = ucb1x00_evt->x;
 			samp->y = ucb1x00_evt->y;
 			samp->pressure = ucb1x00_evt->pressure;
